Phase 6 post-phase audit: security hardening, perf optimization, robustness
Security (Critical):
- glTF loader: bounds-check all accessor/bufferView/buffer indices before access
- glTF loader: validate buffer offset arithmetic prevents out-of-bounds reads
- glTF loader: sanitize texture URIs with path canonicalization (prevent traversal)
- glTF loader: validate scene index against scenes array size
- glTF loader: replace reinterpret_cast with memcpy helpers (strict aliasing)
- OBJ loader: reject files > 256 MB to prevent OOM

Performance (High):
- Eliminate per-frame heap allocations in render pipeline:
  - buildInstanceBatches: pooled member batches with high-water-mark reuse
  - Shadow caster vectors converted to persistent members
  - Cascade culled casters vector reused across cascades
  - Material groups map retains bucket memory across frames
  - selectShadowCastingPointLights fills member instead of returning by value
- Pre-built static uniform name strings for cascade/point shadow uniforms
  (eliminates ~10 std::string temporaries per frame)

Robustness:
- PMR arena: replace raw new/delete with monotonic_buffer_resource::release()
- CpuProfileScope: add missing explicit on single-arg constructor
- Framebuffer/Shader move ctors already had noexcept (verified)
- Water gridResolution already clamped to min 2 (verified)

Code quality:
- terrain.h: fix -Wconversion with explicit static_cast<float>
- aabb.h: rename shadowed variable 'transformed' to 'xformed'
- logger.cpp: replace std::endl with '\n'
- .gitignore: add .cache/ directory

Adds Phase 6 audit report: docs/PHASE6_AUDIT_REPORT.md

+++ engine/environment/terrain.h
@@ -168,10 +168,10 @@ public:
     int getDepth() const { return m_config.depth; }
 
     /// @brief Gets the total world-space terrain size (X extent).
-    float getWorldWidth() const { return (m_config.width - 1) * m_config.spacingX; }
+    float getWorldWidth() const { return static_cast<float>(m_config.width - 1) * m_config.spacingX; }
 
     /// @brief Gets the total world-space terrain size (Z extent).
-    float getWorldDepth() const { return (m_config.depth - 1) * m_config.spacingZ; }
+    float getWorldDepth() const { return static_cast<float>(m_config.depth - 1) * m_config.spacingZ; }
 
     /// @brief Gets the LOD distance ranges for per-vertex morphing.
     const std::vector<float>& getLodRanges() const { return m_lodRanges; }
